requirements.txt
fastapi
uvicorn
pandas
scikit-learn
google-generativeai
python-dotenv
kagglehub
